+++ include/methods.h
@@ -1,0 +1,2 @@
+#include "../src/driveset.c"
+#include "../src/mechanismset.c"
