+++ include/ports.h
@@ -1,0 +1,12 @@
+#define lMotor1 3
+#define lMotor2 4
+
+#define rMotor1 2
+#define rMotor2 5
+
+#define intake1 6
+
+#define conveyor1 1
+
+#define flyWheel1 7
+#define flyWheel2 8
