+++ src/driveset.c
@@ -1,0 +1,11 @@
+#include "main.h"
+#include "ports.h"
+
+void lDriveSet(int control) {
+	motorSet(lMotor1, -control);
+	motorSet(lMotor2, -control);
+}
+void rDriveSet(int control) {
+	motorSet(rMotor1, -control);
+	motorSet(rMotor2, -control);
+}
